Initial work on a SPEAD->HDF5 writer
It involves a lot of duplicated code from other places. The idea is that
eventually the pieces should migrate to other repositories e.g.
telescope_model to kattelmod, file_writer to katdal, rts_model should be
superseded, the script can probably move to katsdpdata...

Sensors still need to be added