Include previous sensor value before start of data
This is important for event sensors that may change infrequently or not
at all during the observation. The last known value for each sensor is
always passed into the telescope state and might be the only value.
The corresponding timestamp could be hours before the start of the
observation and will be missed by get_range() unless the include_previous
flag is enabled.