Set chunk size for filewriter
This should prevent the high disk write load in the filewriter. The
automatic selection puts 16 dumps in a chunk (for 16A/32K) which causes
each chunk to be written to disk 16 times due to flushing. The new
choice puts each dump in its own chunk. It also keeps all baselines, and
real+imag in one chunk, while subdividing the channels more finely.