Reset dumps and rate sensor on capture-init
Previously the dumps sensor would go directly from the total count in
the previous capture session to 1 when the first heap arrives. Now it
goes to 0 on capture-init. Similarly, the rate sensor is set to UNKNOWN
when the capture is started (since there is insufficient data).

Over-the-shoulder reviewed by @ludwigschwardt.