Properly compute the heap size
Instead of hardcoding a memory size to reserve in the socket buffer and
the memory pool, compute a size using data retrieved from the telescope
state.

One catch is that if cam2telstate falls over and never marks itself as
ready (or similarly, if ingest falls over and never sets
sdp_l0_bls_ordering), then ?capture-done will block forever because
there is no way to interrupt wait_key. That can't easily be fixed
without aioredis (and hence Python 3). An option might be to have a
(fairly long) timeout on wait_key, but I'd like to avoid timeout hell,
and anyway a better fix is probably to have master controller detect the
jammed service and restart it.