Support for receiving from multiple ingests
This currently depends on the telstate-metadata-out branch of ingest,
both to set needed metadata and to keep the heap IDs from colliding.

It sets an HDF5 "stream_name" attribute on the /Data group, which I plan
to plumb in to katdal for it know which telstate keys to load with
metadata. That will make it practical to hook up a filewriter to the
continuum L0 stream and get sensible results when opening it with
katdal.

It also renames the command-line arguments for the L0 streams to remove
'spectral' from the name, to reflect the fact that it can now be used to
capture either the spectral or continuum stream. This will require some
changes in katsdpcontroller to use the new argument names.