Protect against multiple ingests getting too out of sync
If one ingest falls more than a dump period behind another, we should
discard its data, rather than having a file in which timestamps jump
back and forth (potentially including duplicated timestamps, each with
part of the band for that timestamp).

If out-of-sync ingests turn out to be a problem in practice, we will
need to add some buffering to filewriter - but I'd rather do that only
if needed, because it will add complexity.