Handle bls_ordering containing Unicode instead of byte strings
This is caused by the Python 3 branch of katcbfsim. It will handle a
straight-up array of strings, but more complex data types with strings
aren't currently handled. It's also not handled for sensors, but I don't
think there is anything in the telescope model that has sensors where
each value is an array of strings.

Note that /TelescopeState is unaffected, because it is all
pickle-encoded.