Wait for all ingests to send stop heaps before stopping
If the ingests aren't quite running in lock-step, then previously the
stream would be stopped the moment the first one sent its stop heap,
leading to data loss. Now fixed to wait for all the stop heaps.

This adds a dependency on spead2 1.5.0, so it won't actually build until
katsdpdockerbase gets updated.