First version of basic vis_writer (untested)
This strips out all references to files and disks and writes the
datasets in a heap to the object store via dask, in the style of
h5toobj.py in katdal. Only the RADOS chunk store is supported for now.

A significant change from h5toobj is the support for substreams that
contain a subset of the channels. The timestamps are still written
at the end of the capture (to the object store), along with the dataset
metadata (to telstate).