Use stream dump_index to handle missing dumps
The dump_index item in the L0 SPEAD stream is a more reliable indicator
of the current dump than the number of received timestamps (especially
if there are more than one vis writer in the future). Use the int_time
to synthesise any missing timestamps, while the chunk store will handle
any missing data chunks for a dump. Store int_time and sync_time on the
object, so that they can be verified before capture starts as well.