Rework status sensor and build info
The description of the status sensor in the docstring was outdated.
Let 'finalising' now refer to the final writing of chunk metadata to
telstate and timestamps to Ceph. Let 'complete' be when things are
really complete. Make missing timestamps an 'error'. Also get the
build info from the katversion string of the package.