Change telstate format of chunk info
For starters, call 'correlator_data' an array and not a dataset
(the former fits in with dask / xarray / katdal and the latter is a
leftover of HDF5). Now, instead of creating a telstate entry for each
array's chunk info, put it into one `chunk_info` dict. This is more
convenient when constructing a VisFlagsWeights from it. Also store
the object store name of the *dataset* as this might not match the
current simplistic capture-block/stream scheme. And make the default
chunk size bigger in anticipation of S3.