Simplify vis_writer around capture_stream
The current plan is to have two S3 buckets per capture block ID,
called <capture block> and <capture_block>_<stream>. The former will
contain metadata, obs reports and other user products, while the latter
is the entry point for the chunk store. This suggests that the second
bucket name should actually be an entity of itself, innovatively
called <capture_stream>. By using an underscore separator instead of
a slash, this ensures that it becomes an S3 bucket name. It also matches
the telstate view to get to the chunk info.

Get rid of obj_base_name for now, as this will interfere with the
bucket naming in the S3 store. It is always possible to return it as
a store-specific prefix later on. There is also no need to store the
"chunk_name" as it is defined to be <capture_stream>. Store a basic
S3 endpoint for starters, just to get going.