Add chunk_name back in as it simplifies katdal
Even though the chunk_name matches the capture_stream telstate prefix,
it is still useful to store it explicitly. This way, telstate can be
prepared with the appropriate views and passed to katdal without also
specifying the capture_stream. When the chunk store needs to find the
dataset it only needs to read 'chunk_name'. Additionally, this allows
the chunk store name of the dataset to change in future without
disruptions.