New v3.9 HDF5 file format
This is a stepping stone towards v4 format, hence the silly number.
The pending update to katcorelib (PR #257) changes obs_params from a
sensor of 'key repr(value)' strings to a proper dict-valued attribute,
and moves all obs attributes and sensors to a capture block namespace.

Use the capture block ID passed to capture_init to name the HDF5 file,
instead of a local timestamp. Generate the timestamp anyway instead of
inferring it from the CBID, as the telstate sensor extraction needs it.

To minimise changes to katdal, pass a telstate with the appropriate
view to the telescope model machinery and handle these obs changes in
TelescopeModel. While we are at it, remove TelescopeModel components
and sensors that are not supported anymore ('dig_overflow',
'mxxx_rx_serial_number', 'mxxx_dig_noise_diode' and 'mxxx_description').