Fix up obs_params for v3.9
The first attempt failed because HDF5 files cannot have objects like
dicts as attribute values. The revised approach leaves obs_params in
telstate but stores the capture_block_id in TelescopeModel instead.
This can be used by katdal to find obs_params. The other obs sensors
are left as is.

Restore the SDP component to access the capture_block_id attribute.
Remove Digitiser as it is not used (forgot to cull it in a previous
commit). Remove params from Observation component to avoid a crash.