Move capture_block_id to top-level attrs
Instead of getting the capture block ID from the `sdp_capture_block_id`
sensor as part of the TelescopeModel extraction during capture-done,
save it as a top-level attribute directly after opening the HDF5 file
in capture-init. This is safer, since capture-init knows this ID
without a doubt, and also gets rid of the reintroduced SDP component
in TelescopeModel (easy come easy go). This also follows the approach
of `stream_name` (which is associated with the data component instead).