Improve v3.9 observation script log
Move the capture block start time back to the first script log entry
to get the full obs script log into the data file. This is where
`capture_init` used to happen in the `CaptureSession` in the days
before v3.9 format anyway. The script log has started up at
`capture_init` but is not available as it is waiting for `capture_init`
to return the capture_block_id. Since the start time is only used
in `capture_done`, fiddle with it there.