Remove all timestamp output from vis_writer
The timestamps can now be synthesised from the dump indices, given
the sync time, integration time and first timestamp that are all
available in telstate (see katsdpingest PR #229). The dump indices
also arrive in the SPEAD stream since katsdpingest PR #220.

Therefore, don't bother synthesising timestamps in vis_writer and
writing them explicitly to the chunk store. It is a chunk with a silly
size and also critical for opening a data set. It makes much more sense
to keep it within telstate and thereby avoid accessing the chunk store
for simple things.