Remove dask for writing data
It was causing some sort of cyclic garbage that was causing heaps to not
be returned to the memory pool. I suspect it was also periodically
thrashing the system (lots of lots of threads) which may have
contributed to data being dropped as the network thread may have been
starved. It was only really needed for the network stores (which had
high latency), and since we're now using a local store I've removed it.
On my laptop I still get >1GB/s write speeds (presumably that's into the
page cache rather than true disk writes).

This is a hack-and-slash job on the dask removal for testing. If
successful, it will be cleaned up and the logs made less noisy.

This will hopefully address SR-1100, and also reduce memory usage.