Rip out HDF5 filewriter
Only the MVFv4 vis_writer remains. This allows a lot of the code to be
deleted and dependencies removed.

Some of the deleted code is still used by katsdpbfingest. It will be
moved there in another PR.

The package name is now a bit of a misnomer, but I'll leave that for
another time since it'll have ramifications in other packages.