Fix bug when accessing prefix in flag stream
Make the prefix a proper attribute as opposed to a private one, since
we need the prefix before the first data can be written to the chunk
store (and chunk info is empty at that stage). Test for empty dicts
instead of using ints and None as proxies. Update docstrings.