Unify the code for writing 'complete' files
The os.sync that was in the vis_writer version has been removed. I'd
originally put it there based on a flawed understanding of what the
complete marker meant i.e. that the vis_trawler took it to mean the data
was definitely safely on disk and could transfer it even if the system
had crashed and rebooted. In practice, the data is transferred as it is
written, so an os.sync here would at best ensure durability of a few
chunks, and it could also block the event loop for a significant time.