Extra metrics, avoid blocking event loop
There are three changes:

1. The mechanism to bound the work passed to the executor has changed
from using threading.BoundedSemaphore in the executor to using
asyncio.BoundedSemaphore in ChunkStoreRechunker. Using asyncio to block
rather than threading means that the event loop is no longer completely
blocked (so katcp commands can still be processed for example).

2. A new active-chunks sensor is added to track the number of chunks
that have been posted to the executor but not completed.

3. A bokeh-based dashboard is added to track the sensors at 10Hz (giving
much more resolution than Prometheus). It buffers up to 20 minutes of data,
and by default displays the last 2 minutes.