Handle underscore or dash. Add Prefix field
In order to seemlessly handle the change to dashes, we will maintain the
CaptureStreamId as is, so it means the same thing for both types of
observations. However now we will include a Prefix which is provided by
katdal which provides the bucket holding the streams data.

This will require a change in katdal to provide the prefix.