Support dumps with multiple chunks
The number of dumps per chunk starts at 1, and is doubled until either
the max object size or a maximum dumps per chunk is reached (the latter
being useful both to limit data loss in the event of a crash, and to
constrain memory usage).

Also add --obj-max-channels option to bound the number of channels per
chunk, for spectral imaging.