SPR1-2841 Remove test bypass so that it can fail
This removes the check that bypasses the entire met_extractor test
instead of properly skipping it (and it is not even needed here).

Now the test properly fails as it should with the latest katdal.

Also skip the prod_handler tests if the test data is unavailable
to avoid duping the user.