SPR1-2841 Fix katdal name API usage
The `katdata.source.data` member has no `name` attribute anymore.

Use the top-level name of the dataset, as this now corresponds to the
capture stream ID since katdal 0.19 (see SPR1-1134 for more info).

Don't look up the CBID and stream in `katdata.sources.metadata.attrs`.
These have turned into source attributes for convenience.

While we are at it, fix a FlagProduct bug. The extractor never had
a `_katdata` attribute to begin with.

This improves the coverage of `met_extractors.py` from 20% to 54%.