SPR1-2841 Add many more metadata checks
Check that the extracted metadata matches the test dataset contents.

I still wonder if we should fix 'Targets' (remove the aliases) and
'ElAz' (store it for radec targets too if source motion is OK).

The coverage is still stuck at 54%. The problem is that the products
besides `MeerKATTelescopeProduct` are deprecated as far as I can tell.